+++ core/network/src/main/java/com/hilingual/core/network/di/qualifier/Qualifiers.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.core.network.di
+package com.hilingual.core.network.di.qualifier
 
 import javax.inject.Qualifier
