+++ data/auth/src/main/java/com/hilingual/data/auth/datasourceimpl/AuthLocalDataSourceImpl.kt
@@ -4,7 +4,7 @@ import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
-import com.hilingual.data.auth.datasource.AuthDataStore
+import com.hilingual.data.auth.di.quilifier.AuthDataStore
 import com.hilingual.data.auth.datasource.AuthLocalDataSource
 import com.hilingual.data.auth.localstorage.model.TokenPreferences
 import com.hilingual.core.localstorage.di.qualifier.UserInfoDataStore
