+++ data/auth/src/main/java/com/hilingual/data/auth/di/AuthDataStoreModule.kt
@@ -3,7 +3,7 @@ package com.hilingual.data.auth.di
 import android.content.Context
 import androidx.datastore.core.DataStore
 import com.hilingual.core.localstorage.builder.DataStoreBuilder
-import com.hilingual.data.auth.datasource.AuthDataStore
+import com.hilingual.data.auth.di.quilifier.AuthDataStore
 import com.hilingual.data.auth.localstorage.model.TokenPreferences
 import dagger.Module
 import dagger.Provides
